+++ packages/tailwindcss/src/utilities.ts
@@ -254,7 +254,7 @@ function resolveThemeColor<T extends ThemeKey>(
       break
     }
     case 'current': {
-      value = 'currentColor'
+      value = 'currentcolor'
       break
     }
     default: {
@@ -4298,7 +4298,8 @@ export function createUtilities(theme: Theme) {
 
       if (!candidate.value) {
         let value = theme.get(['--drop-shadow'])
-        if (value === null) return
+        let resolved = theme.resolve(null, ['--drop-shadow'])
+        if (value === null || resolved === null) return
 
         return [
           filterProperties(),
@@ -4309,7 +4310,12 @@ export function createUtilities(theme: Theme) {
             alpha,
             (color) => `var(--tw-drop-shadow-color, ${color})`,
           ),
-          decl('--tw-drop-shadow', `drop-shadow(${theme.resolve(null, ['--drop-shadow'])})`),
+          decl(
+            '--tw-drop-shadow',
+            segment(resolved, ',')
+              .map((value) => `drop-shadow(${value})`)
+              .join(' '),
+          ),
           decl('filter', cssFilterValue),
         ]
       }
@@ -4350,7 +4356,8 @@ export function createUtilities(theme: Theme) {
       // Shadow size
       {
         let value = theme.get([`--drop-shadow-${candidate.value.value}`])
-        if (value) {
+        let resolved = theme.resolve(candidate.value.value, ['--drop-shadow'])
+        if (value && resolved) {
           if (candidate.modifier && !alpha) return
 
           if (alpha) {
@@ -4379,7 +4386,9 @@ export function createUtilities(theme: Theme) {
             ),
             decl(
               '--tw-drop-shadow',
-              `drop-shadow(${theme.resolve(candidate.value.value, ['--drop-shadow'])})`,
+              segment(resolved, ',')
+                .map((value) => `drop-shadow(${value})`)
+                .join(' '),
             ),
             decl('filter', cssFilterValue),
           ]
@@ -5439,7 +5448,7 @@ export function createUtilities(theme: Theme) {
 
     staticUtility('ring-inset', [boxShadowProperties, ['--tw-ring-inset', 'inset']])
 
-    let defaultRingColor = theme.get(['--default-ring-color']) ?? 'currentColor'
+    let defaultRingColor = theme.get(['--default-ring-color']) ?? 'currentcolor'
     function ringShadowValue(value: string) {
       return `var(--tw-ring-inset,) 0 0 0 calc(${value} + var(--tw-ring-offset-width)) var(--tw-ring-color, ${defaultRingColor})`
     }
@@ -5515,7 +5524,7 @@ export function createUtilities(theme: Theme) {
     ])
 
     function insetRingShadowValue(value: string) {
-      return `inset 0 0 0 ${value} var(--tw-inset-ring-color, currentColor)`
+      return `inset 0 0 0 ${value} var(--tw-inset-ring-color, currentcolor)`
     }
     utilities.functional('inset-ring', (candidate) => {
       if (!candidate.value) {
@@ -5683,6 +5692,16 @@ export function createUtilities(theme: Theme) {
   return utilities
 }
 
+// Only allowed bare value data types, to prevent creating new syntax that we
+// typically don't support right now. E.g.: `--value(color)` would allow you to
+// use `text-#0088cc` as a valid utility, which is not what we want.
+export const BARE_VALUE_DATA_TYPES = [
+  'number', // 2.5
+  'integer', // 8
+  'ratio', // 2/3
+  'percentage', // 25%
+]
+
 export function createCssUtility(node: AtRule) {
   let name = node.params
 
@@ -5815,7 +5834,6 @@ export function createCssUtility(node: AtRule) {
           }
           fn.nodes = ValueParser.parse(args.join(','))
 
-          // Track information for suggestions
           for (let node of fn.nodes) {
             // Track literal values
             if (
@@ -5831,6 +5849,36 @@ export function createCssUtility(node: AtRule) {
             else if (node.kind === 'word' && node.value[0] === '-' && node.value[1] === '-') {
               let value = node.value.replace(/-\*.*$/g, '') as `--${string}`
               storage[fn.value].themeKeys.add(value)
+            }
+
+            // Validate bare value data types
+            else if (
+              node.kind === 'word' &&
+              !(node.value[0] === '[' && node.value[node.value.length - 1] === ']') && // Ignore arbitrary values
+              !BARE_VALUE_DATA_TYPES.includes(node.value)
+            ) {
+              console.warn(
+                `Unsupported bare value data type: "${node.value}".\nOnly valid data types are: ${BARE_VALUE_DATA_TYPES.map((x) => `"${x}"`).join(', ')}.\n`,
+              )
+              // TODO: Once we properly track the location of the node, we can
+              //       clean this up in a better way.
+              let dataType = node.value
+              let copy = structuredClone(fn)
+              let sentinelValue = '¶'
+              ValueParser.walk(copy.nodes, (node, { replaceWith }) => {
+                if (node.kind === 'word' && node.value === dataType) {
+                  replaceWith({ kind: 'word', value: sentinelValue })
+                }
+              })
+              let underline = '^'.repeat(ValueParser.toCss([node]).length)
+              let offset = ValueParser.toCss([copy]).indexOf(sentinelValue)
+              let output = [
+                '```css',
+                ValueParser.toCss([fn]),
+                ' '.repeat(offset) + underline,
+                '```',
+              ].join('\n')
+              console.warn(output)
             }
           }
         })
@@ -6075,12 +6123,7 @@ function resolveValueFunction(
       // Limit the bare value types, to prevent new syntax that we
       // don't want to support. E.g.: `text-#000` is something we
       // don't want to support, but could be built this way.
-      if (
-        arg.value !== 'number' &&
-        arg.value !== 'integer' &&
-        arg.value !== 'ratio' &&
-        arg.value !== 'percentage'
-      ) {
+      if (!BARE_VALUE_DATA_TYPES.includes(arg.value)) {
         continue
       }
 
@@ -6183,13 +6226,20 @@ function alphaReplacedShadowProperties(
     return varInjector(replaceAlpha(color, alpha))
   })
 
+  function applyPrefix(x: string) {
+    if (!prefix) return x
+    return segment(x, ',')
+      .map((value) => prefix + value)
+      .join(',')
+  }
+
   if (requiresFallback) {
     return [
-      decl(property, prefix + replaceShadowColors(value, varInjector)),
-      rule('@supports (color: lab(from red l a b))', [decl(property, prefix + replacedValue)]),
+      decl(property, applyPrefix(replaceShadowColors(value, varInjector))),
+      rule('@supports (color: lab(from red l a b))', [decl(property, applyPrefix(replacedValue))]),
     ]
   } else {
-    return [decl(property, prefix + replacedValue)]
+    return [decl(property, applyPrefix(replacedValue))]
   }
 }